requirements.txt
aiohttp==3.8.1
python-dotenv==0.19.1
schedule==1.1.0
pytest==7.3.1
pytest-asyncio==0.21.0